pyproject.toml
[tool.poetry]
name = "push-notification-service"
version = "0.1.0"
description = "Microservice for managing push notification tokens"
authors = ["OpenHands <openhands@all-hands.dev>"]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.5"
uvicorn = "^0.32.1"
sqlalchemy = "^2.0.36"
asyncpg = "^0.30.0"
alembic = "^1.14.0"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.10"

[tool.poetry.dev-dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.23.5"
httpx = "^0.27.0"
black = "^24.2.0"
isort = "^5.13.2"
flake8 = "^7.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py312']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3